+++ extensions/wrapper/wrapper-common-api/src/main/java/de/sovity/edc/ext/wrapper/api/common/model/PermissionDto.java
@@ -40,5 +40,5 @@ public class PermissionDto {
 
     @Schema(description = "Possible constraints for the permission",
             requiredMode = RequiredMode.REQUIRED)
-    private ExpressionDto constraints;
+    private Expression expression;
 }