+++ extensions/wrapper/wrapper-common-mappers/src/test/java/de/sovity/edc/ext/wrapper/api/common/mappers/PolicyMapperTest.java
@@ -3,6 +3,9 @@ package de.sovity.edc.ext.wrapper.api.common.mappers;
 import de.sovity.edc.ext.wrapper.api.common.mappers.policy.AtomicConstraintMapper;
 import de.sovity.edc.ext.wrapper.api.common.mappers.policy.ConstraintExtractor;
 import de.sovity.edc.ext.wrapper.api.common.mappers.policy.MappingErrors;
+import de.sovity.edc.ext.wrapper.api.common.model.AtomicConstraintDto;
+import de.sovity.edc.ext.wrapper.api.common.model.Expression;
+import de.sovity.edc.ext.wrapper.api.common.model.ExpressionType;
 import de.sovity.edc.ext.wrapper.api.common.model.UiPolicyConstraint;
 import de.sovity.edc.ext.wrapper.api.common.model.UiPolicyCreateRequest;
 import jakarta.json.JsonObject;
@@ -14,6 +17,8 @@ import org.eclipse.edc.spi.result.Result;
 import org.eclipse.edc.transform.spi.TypeTransformerRegistry;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockedStatic;
@@ -21,8 +26,10 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.util.List;
 
+import static de.sovity.edc.ext.wrapper.api.common.model.ExpressionType.ATOMIC_CONSTRAINT;
 import static de.sovity.edc.utils.JsonUtils.parseJsonObj;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.mockStatic;
@@ -87,5 +94,33 @@ class PolicyMapperTest {
         assertThat(actual.getPermissions().get(0).getConstraints()).hasSize(1);
         assertThat(actual.getPermissions().get(0).getAction().getType()).isEqualTo("USE");
         assertThat(actual.getPermissions().get(0).getConstraints().get(0)).isSameAs(expected);
+    }
+
+    @ParameterizedTest
+    @ValueSource(strings = {"AND", "OR", "XOR"})
+    void buildGenericPolicy(String constraintTypeString) {
+        // arrange
+        var expressionType = ExpressionType.valueOf(constraintTypeString);
+        var incomingConstraint = mock(AtomicConstraintDto.class);
+        var mappedAtomicConstraint = mock(AtomicConstraint.class);
+        var atomicConstraint = new Expression(ATOMIC_CONSTRAINT, List.of(), incomingConstraint);
+        var atomicConstraints = List.of(atomicConstraint, atomicConstraint);
+        var baseConstraintElement = new Expression(expressionType, atomicConstraints, null);
+
+        // act
+        when(atomicConstraintMapper
+                .buildAtomicConstraint(eq(incomingConstraint)))
+                .thenReturn(mappedAtomicConstraint);
+        var policy = policyMapper.buildPolicy(List.of(baseConstraintElement));
+
+        // assert
+        assertThat(policy.getType()).isEqualTo(PolicyType.SET);
+        assertThat(policy.getPermissions()).hasSize(1);
+        var permission = policy.getPermissions().get(0);
+        assertThat(permission.getConstraints()).hasSize(1);
+        assertThat(permission.getAction().getType()).isEqualTo("USE");
+
+        var constraintObject = permission.getConstraints().get(0);
+        assertNotNull(constraintObject);
     }
 }