+++ extensions/wrapper/wrapper/src/main/java/de/sovity/edc/ext/wrapper/WrapperExtensionContextBuilder.java
@@ -278,7 +278,8 @@ public class WrapperExtensionContextBuilder {
         var useCaseResource = new UseCaseResourceImpl(
                 kpiApiService,
                 supportedPolicyApiService,
-                useCaseCatalogApiService
+                useCaseCatalogApiService,
+                policyDefinitionApiService
         );
 
         // Collect all JAX-RS resources